Enums and Pattern matching
So first I learned to use enum, add #[derive(Debug)] to allow printing that enum,
impl block for an enum, attaching a data with the enum, match case, Option<t> enum
to avoid the NULL value problem.

option<t> wraps the null value so that we always need to check before using the value.

Then the last part was to learn if let statement to use single match i.e. without the
exhaustive search.